+++ inst/include/dust/gpu/filter_kernels.hpp
@@ -2,7 +2,7 @@
 #define DUST_CUDA_FILTER_KERNELS_HPP
 
 #include <cmath>
-#include "dust/cuda/cuda.hpp"
+#include "dust/gpu/cuda.hpp"
 
 namespace dust {
