+++ inst/include/dust/gpu/filter_state.hpp
@@ -2,7 +2,7 @@
 #define DUST_CUDA_FILTER_STATE_HPP
 
 #include "dust/filter_state.hpp"
-#include "dust/cuda/device_state.hpp"
+#include "dust/gpu/device_state.hpp"
 
 namespace dust {
 namespace filter {
