+++ inst/include/dust/gpu/kernels.hpp
@@ -1,7 +1,7 @@
 #ifndef DUST_CUDA_KERNELS_HPP
 #define DUST_CUDA_KERNELS_HPP
 
-#include "dust/cuda/device_state.hpp"
+#include "dust/gpu/device_state.hpp"
 #include "dust/utils.hpp"
 
 namespace dust {
