+++ inst/include/dust/interface/gpu.hpp
@@ -8,8 +8,8 @@
 #include <cpp11/logicals.hpp>
 #include <cpp11/strings.hpp>
 
-#include "dust/cuda/launch_control.hpp"
-#include "dust/interface/cuda_gpu_info.hpp"
+#include "dust/gpu/launch_control.hpp"
+#include "dust/interface/gpu_info.hpp"
 #include "dust/interface/helpers.hpp"
 
 namespace dust {
